refactor: separate footer component
